@@ -3,7 +3,7 @@
 #include "area.h"
 #include "engine/color.h"
 #include "engine/geometry.h"
-#include <boost/unordered_map.hpp>
+#include <unordered_map>
 #include <fstream>
 #include <functional>
 #include <memory>
@@ -38,8 +38,8 @@ private:
     static Vector2 globalPositionToTilePosition(Vector2 position);
 
     const Game* game;
-    boost::unordered_map<Vector3, Area> areas;
-    boost::unordered_map<Vector3, int64_t> savedAreaOffsets;
+    std::unordered_map<Vector3, Area> areas;
+    std::unordered_map<Vector3, int64_t> savedAreaOffsets;
     std::unique_ptr<SaveFile> saveFile;
     Color32 sunlight;
 };
