@@ -3,7 +3,7 @@
 #include "creature.h"
 #include <stdexcept>
 
-std::unique_ptr<AI> AI::get(boost::string_ref id, Creature& creature)
+std::unique_ptr<AI> AI::get(std::string_view id, Creature& creature)
 {
     if (id == "AttackNearestEnemy") return std::make_unique<AttackNearestEnemy>(creature);
 
