@@ -5,7 +5,7 @@
 
 Component::~Component() {}
 
-std::unique_ptr<Component> Component::get(boost::string_ref name, Entity& parent)
+std::unique_ptr<Component> Component::get(std::string_view name, Entity& parent)
 {
     std::unique_ptr<Component> component;
 
