@@ -2,7 +2,7 @@
 
 #include "component.h"
 #include "engine/config.h"
-#include <boost/utility/string_ref.hpp>
+#include <string_view>
 #include <memory>
 #include <string>
 #include <vector>
@@ -10,7 +10,7 @@
 class Entity
 {
 public:
-    Entity(boost::string_ref id, const Config& config);
+    Entity(std::string_view id, const Config& config);
     Entity(const Entity&) = delete;
     Entity(Entity&&) = default;
     Entity& operator=(const Entity&) = delete;
@@ -19,7 +19,7 @@ public:
 
     std::string getName() const;
     std::string getNameIndefinite() const;
-    boost::string_ref getId() const { return id; }
+    std::string_view getId() const { return id; }
     const Config& getConfig() const { return *config; }
     template<typename ComponentType>
     std::vector<ComponentType*> getComponentsOfType() const;
