@@ -4,11 +4,9 @@
 #include "gui.h"
 #include "tile.h"
 #include "engine/savefile.h"
-#include <boost/algorithm/string/erase.hpp>
-#include <boost/algorithm/string/predicate.hpp>
 #include <iostream>
 
-static Color16 getMaterialColor(boost::string_ref materialId)
+static Color16 getMaterialColor(std::string_view materialId)
 {
     if (!materialId.empty())
     {
@@ -28,15 +26,15 @@ static Color16 getMaterialColor(boost::string_ref materialId)
         return Color16::none;
 }
 
-Item::Item(boost::string_ref id, boost::string_ref materialId)
+Item::Item(std::string_view id, std::string_view materialId)
 :   Item(id, materialId,
          ::getSprite(*Game::itemSpriteSheet, *Game::itemConfig, id, 0, getMaterialColor(materialId)))
 {
 }
 
-Item::Item(boost::string_ref id, boost::string_ref materialId, Sprite sprite)
+Item::Item(std::string_view id, std::string_view materialId, Sprite sprite)
 :   Entity(id, *Game::itemConfig),
-    materialId(materialId.to_string()),
+    materialId(materialId),
     sprite(std::move(sprite))
 {
 }
@@ -45,13 +43,14 @@ std::unique_ptr<Item> Item::load(const SaveFile& file)
 {
     auto itemId = file.readString();
     std::unique_ptr<Item> item;
+    std::string_view corpseSuffix = "Corpse";
 
-    if (boost::algorithm::ends_with(itemId, "Corpse"))
+    if (endsWith(itemId, corpseSuffix))
     {
         if (file.readBool())
             item = std::make_unique<Corpse>(std::make_unique<Creature>(file, nullptr));
         else
-            item = std::make_unique<Corpse>(boost::algorithm::erase_last_copy(itemId, "Corpse"));
+            item = std::make_unique<Corpse>(removeSuffix(itemId, corpseSuffix));
     }
     else
     {
@@ -98,12 +97,12 @@ bool Item::use(Creature& user, Game& game)
 
 bool Item::isEdible() const
 {
-    return Game::itemConfig->getOptional<bool>(getId(), "isEdible").get_value_or(false);
+    return Game::itemConfig->getOptional<bool>(getId(), "isEdible").value_or(false);
 }
 
 EquipmentSlot Item::getEquipmentSlot() const
 {
-    auto slotString = getConfig().getOptional<std::string>(getId(), "EquipmentSlot").get_value_or("Hand");
+    auto slotString = getConfig().getOptional<std::string>(getId(), "EquipmentSlot").value_or("Hand");
 
     if (slotString == "Head") return Head;
     if (slotString == "Torso") return Torso;
@@ -131,7 +130,7 @@ void Item::renderEquipped(Window& window, Vector2 position) const
     sprite.render(window, position, equippedSourceOffset);
 }
 
-std::string getRandomMaterialId(boost::string_ref itemId)
+std::string getRandomMaterialId(std::string_view itemId)
 {
     auto materials = Game::itemConfig->get<std::vector<std::string>>(itemId, "PossibleMaterials");
     return materials.empty() ? "" : randomElement(materials);
@@ -145,7 +144,7 @@ Corpse::Corpse(std::unique_ptr<Creature> creature)
     sprite.setAsciiGlyph(corpseGlyph);
 }
 
-Corpse::Corpse(boost::string_ref creatureId)
+Corpse::Corpse(std::string_view creatureId)
 :   Item(creatureId + "Corpse", "", ::getSprite(*Game::creatureSpriteSheet, *Game::creatureConfig,
                                                 creatureId, corpseFrame, Color32::none))
 {
