@@ -3,7 +3,7 @@
 #include "entity.h"
 #include "engine/geometry.h"
 #include "engine/sprite.h"
-#include <boost/utility/string_ref.hpp>
+#include <string_view>
 #include <memory>
 
 class Creature;
@@ -14,7 +14,7 @@ enum EquipmentSlot : int;
 class Item : public Entity
 {
 public:
-    Item(boost::string_ref id, boost::string_ref materialId);
+    Item(std::string_view id, std::string_view materialId);
     static std::unique_ptr<Item> load(const SaveFile& file);
     virtual void save(SaveFile& file) const;
     virtual void exist() {}
@@ -28,19 +28,19 @@ public:
     const Sprite& getSprite() const { return sprite; }
 
 protected:
-    Item(boost::string_ref id, boost::string_ref materialId, Sprite sprite);
+    Item(std::string_view id, std::string_view materialId, Sprite sprite);
 
     std::string materialId;
     Sprite sprite;
 };
 
-std::string getRandomMaterialId(boost::string_ref itemId);
+std::string getRandomMaterialId(std::string_view itemId);
 
 class Corpse final : public Item
 {
 public:
     Corpse(std::unique_ptr<Creature> creature);
-    Corpse(boost::string_ref creatureId);
+    Corpse(std::string_view creatureId);
     void exist() override;
     void renderEquipped(Window& window, Vector2 position) const override;
     void save(SaveFile& file) const override;
