@@ -3,7 +3,7 @@
 #include "gui.h"
 #include "engine/font.h"
 #include "engine/savefile.h"
-#include <boost/utility/string_ref.hpp>
+#include <string_view>
 #include <deque>
 #include <string>
 
@@ -13,7 +13,7 @@ class Message
 {
 public:
     Message(std::string&& text, int turn) : text(std::move(text)), turn(turn), count(1) {}
-    boost::string_ref getText() const { return text; }
+    std::string_view getText() const { return text; }
     int getTurn() const { return turn; }
     int getCount() const { return count; }
     void increaseCount(int currentTurn) { ++count; turn = currentTurn; }
@@ -32,7 +32,7 @@ namespace MessageSystem
                       int currentTurn);
 
 #ifdef DEBUG
-    void addDebugMessage(boost::string_ref message, MessageType = Normal);
+    void addDebugMessage(std::string_view message, MessageType = Normal);
     void addToCommandHistory(std::string&& command);
     std::string getPreviousCommand();
     std::string getNextCommand();
